Added SQLServerLogonConnectionTrigger.sql
Add the SQL Server logon connection trigger. This trigger should monitor all user logons to a SQL Server instance and log it.
It is supposed to reveal who is gaining access to SQL Server and what applications they're using even though the application name field can say anything and have nothing to do with the real application that is making the connection.